aoc2020.Tests/Day09Tests.cs
using System.IO;
using aoc2020.Days;
using Xunit;

namespace aoc2020.Tests
{
    public class Day09Tests
    {
        [Fact]
        public void TestPart1()
        {
            var assertions = new[]
            {
                ("Inputs/09.txt", 127L),
            };
            assertions.ForEach(((string Input, long Expected) x) =>
                Assert.Equal(x.Expected, Day09.Part1(File.ReadAllText(x.Input).SplitAsLong(), 5)));
        }

        [Fact]
        public void TestPart2()
        {
            var assertions = new[]
            {
                ("Inputs/09.txt", 62L),
            };
            assertions.ForEach(((string Input, long Expected) x) =>
                Assert.Equal(x.Expected, Day09.Part2(File.ReadAllText(x.Input).SplitAsLong(), 5)));
        }
    }
}
